requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
python-dotenv==1.0.0
Flask==3.0.0
Werkzeug==3.0.1
